requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
Werkzeug==3.1.3
selenium==4.18.1
bs4==0.0.2
pandas==2.2.1
openpyxl==3.1.2
xlrd>=2.0.1
google-generativeai
python-dotenv
python-pptx==0.6.23
boto3==1.34.162